# openapi.yml
openapi: 3.0.3
info:
  title: pomu.app API
  version: 0.0.1
  description: Public API for pomu.app
  contact:
    name: General inqueries regarding API
    email: mari@pomu.app
    url: https://pomu.app
  license:
    name: BSD 3-Clause
    url: https://github.com/mellowagain/pomu/blob/main/LICENSE.txt
servers:
  - url: https://pomu.app/api
    description: Production
  - url: https://dev.pomu.app/api
    description: Development instance
components:
  schemas:
    video:
      type: object
      required:
        - id
        - submitters
        - scheduledStart
        - finished
        - title
        - channelName
        - channelId
        - thumbnail
      properties:
        id:
          type: string
          description: Video ID
        submitters:
          type: array
          description: "List of user IDs which have submitted this video to the queue (Format: Provider/UserID)"
          items:
            type: string
        scheduledStart:
          type: string
          format: date-time
          description: Scheduled start of livestream
        finished:
          type: boolean
          description: Whenever the livestream has finished already
        title:
          type: string
          description: Livestream title
        channelName:
          type: string
          description: Name of channel which is streaming this livestream
        channelId:
          type: string
          description: Channel ID of channel which is streaming this livestream
        thumbnail:
          type: string
          description: URL to maximum quality thumbnail available for livestream
        downloadUrl:
          type: string
          description: URL where archive is available for download
        fileSizeBytes:
          type: integer
          format: int64
          description: Archive file size in bytes
        length:
          type: integer
          format: int32
          description: Length of livestream in seconds
...
  parameters:
    videoId:
      name: videoId
      in: path
      required: true
      schema:
        type: string
        description: Video ID of the video which you want to download `type` from
paths:
...
  /qualities:
    get:
      operationId: PeekForQualities
      description: Get list of available qualities for stream
      parameters:
        - name: url
          in: query
          description: YouTube url of livestream for which qualities should be peeked
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  required:
                    - code
                    - resolution
                    - best
                  properties:
                    code:
                      type: integer
                      format: int32
                      description: Numeric ID of the quality
                    resolution:
                      type: string
                      description: Human-readable resolution of the quality
                    best:
                      type: boolean
                      description: Whenever this is the best available quality (both audio and video)
  /submit:
    post:
      operationId: SubmitVideo
      description: Submits a livestream to the archiving queue
      requestBody:
        required: true
        description: Livestream submission payload
        content:
          application/json:
            schema:
              type: object
              required:
                - videoUrl
                - quality
              properties:
                videoUrl:
                  type: string
                  description: YouTube url of live stream which should be added to queue
                quality:
                  type: integer
                  format: int32
                  description: Numeric ID of the quality in which the live stream should be archived
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/video"
        "401":
          description: Not logged in
...
  /history:
    get:
      operationId: GetHistory
      description: Gets list of archived livestreams
      parameters:
        - name: page
          in: query
          description: Page to display
          schema:
            type: integer
            format: int32
            additionalProperties:
              minimum: 0
              default: 0
        - name: limit
          in: query
          description: Amount of livestreams to display per page
          schema:
            type: integer
            format: int32
            additionalProperties:
              maximum: 100
              default: 25
        - name: sort
          in: query
          description: Sort direction of results
          schema:
            type: string
            enum:
              - asc
              - desc
        - name: unfinished
          in: query
          description: Whenever unfinished livestreams (live or upcoming) should be displayed
          schema:
            type: boolean
            additionalProperties:
              default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/video"
          headers:
            X-Pomu-Pagination-Total:
              description: Total amounts of rows that match the filter
              required: true
              schema:
                type: integer
            X-Pomu-Pagination-Has-More:
              description: Whenever there are more pages available
              required: true
              schema:
                type: boolean
...
  /download/{videoId}/{type}:
    parameters:
      - $ref: "#/components/parameters/videoId"
      - name: type
        in: path
        required: true
        schema:
          type: string
          description: What you want to download
          enum:
            - video
            - ffmpeg
            - thumbnail
      - name: User-Agent
        in: header
        required: true
        schema:
          type: string
          description: |
            A identify-able user agent string is **required** to use the pomu download endpoint.
            Including a identify-able user agent allows us to contact you in case of abuse
            or usage outside of our fair-use policy.

            Failing to include such a user agent **will** result in request blocking (401 Bad Request)
            and may also lead to an automated IP(-range) ban, if the request block is being circumvented.
        example: "pomu (https://github.com/mellowagain/pomu)"
        allowReserved: true
    get:
      operationId: Download
      description: Downloads the request `type` of video `videoId`
      responses:
        "307":
          description: | 
            OK - Follow redirect (response header `Location`) to download the requested file.

            > warn
            > The returned `Content-Type` and `Content-Disposition` may differ depending on
            > what type you requested and when the stream was archived:

            - **Type `video`** archived before 2023-01-26:
              - `Content-Type: binary/octet-stream`
              - No `Content-Disposition` header
            - **Type `video`** archived between 2023-01-27 and 2023-02-04:
              - `Content-Type: video/mp4`
              - `Content-Disposition: inline`
            - **Type `video`** archived after 2023-02-05:
              - `Content-Type: video/mp4`
              - `Content-Disposition: attachment`
            - **Type `thumbnail`** archived before 2023-01-26:
              - `Content-Type: binary/octet-stream`
              - No `Content-Disposition` header
            - **Type `thumbnail`** archived after 2023-01-27:
              - `Content-Type: image/jpeg`
              - `Content-Disposition: inline`
            - **Type `ffmpeg`** archived before 2023-01-26:
              - `Content-Type: binary/octet-stream`
              - No `Content-Disposition` header
            - **Type `ffmpeg`** archived after 2023-01-27:
              - `Content-Type: text/plain`
              - `Content-Disposition: inline`
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: | 
            # Bad Request
            One of these two things occurred:
            - Video or ffmpeg log download has been requested but the video has not yet finished archiving.
            - A bad user agent has been sent with the request. Please see the `User-Agent` header definition in this file.
        "404":
          description: Not Found
...